requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
groq==0.11.0
psycopg2-binary==2.9.10
apscheduler==3.10.4
python-dotenv==1.0.1
jinja2==3.1.4
httpx==0.27.2
